WIP truncation
../lib/gen_tables.c:89:22: warning: overflow in conversion from ‘long long unsigned int’ to ‘int’ changes value from ‘4294967296’ to ‘0’ [-Woverflow]
   89 | #define _S(VAL, S) { (VAL), (S), 0, 0 },
      |                      ^
./clone-flagtab.h:47:1: note: in expansion of macro ‘_S’
   47 | _S(0x100000000ULL,      "CLONE_CLEAR_SIGHAND")
      | ^~

+++ auparse/auparse-idata.h
@@ -39,8 +39,8 @@ typedef struct _idata {
 } idata;
 
 
-int auparse_interp_adjust_type(int rtype, const char *name, const char *val);
-char *auparse_do_interpretation(int type, const idata *id,
+long long int auparse_interp_adjust_type(int rtype, const char *name, const char *val);
+char *auparse_do_interpretation(long long int type, const idata *id,
 	auparse_esc_t escape_mode);
 void _auparse_load_interpretations(const char *buf);
 void _auparse_free_interpretations(void);
